Add intRVFL to the models module
